fix: resolve hypothesize clustering and graduate deserialization 500 error
- Update cluster_facts in pipeline.rs to use hybrid Jaccard token overlap similarity when dense vector embeddings are zeroed/missing
- Fix WikiNode.metacognitive_confidence type from Option<i32> to Option<f64> across contracts, pipeline, watcher, ingestion, and backend
- Update graduation_pipeline.rs SQL to cast record ID as string (SELECT *, type::string(id) AS id FROM wiki_node)
- Verify hypothesize queues 6 synthesis tasks from 312 facts and graduate completes cleanly

+++ mythrax-core/src/vault/watcher.rs
@@ -468,7 +468,7 @@ struct WikiFrontmatter {
     name: Option<String>,
     scope: Option<String>,
     edges: Option<Vec<FrontmatterEdge>>,
-    metacognitive_confidence: Option<i32>,
+    metacognitive_confidence: Option<f64>,
     node_type: Option<String>,
     item_type: Option<String>,
 }
